division en varios componentes
