+++ src/components/controlPanel.jsx
@@ -1,0 +1,27 @@
+import React from 'react'
+import { Box, Button } from '@mui/material'
+
+export default function ControlPanel({ onNewGame, onCheckSolution, onSolveBacktracking, onSolveBranchAndBound, onClearSolution, onResetGame }) {
+  return (
+    <Box sx={{ mt: 2, display: 'flex', flexWrap: 'wrap', gap: 2, justifyContent: 'center' }}>
+      <Button variant="contained" onClick={onNewGame}>
+        Nuevo juego
+      </Button>
+      <Button variant="contained" onClick={onCheckSolution}>
+        Verificar
+      </Button>
+      <Button variant="contained" onClick={onSolveBacktracking}>
+        Resolver (Backtracking)
+      </Button>
+      <Button variant="contained" onClick={onSolveBranchAndBound}>
+        Resolver (R&A)
+      </Button>
+      <Button variant="contained" onClick={onClearSolution}>
+        Borrar solución
+      </Button>
+      <Button variant="outlined" onClick={onResetGame}>
+        Nuevo juego / Modo
+      </Button>
+    </Box>
+  )
+}
